Label int to double

+++ src/shared/Label.java
@@ -4,13 +4,13 @@ package shared;
  * Keeps label of a sample to be provided to machin learning algorithm.
  */
 public class Label {
-    private int valence;
-    private int arousal;
-    private int dominance;
-    private int liking;
+    private double valence;
+    private double arousal;
+    private double dominance;
+    private double liking;
     private Emotion emotion;
 
-    public Label(int liking, int valence, int arousal, int dominance, Emotion emotion) {
+    public Label(double liking, double valence, double arousal, double dominance, Emotion emotion) {
         this.liking = liking;
         this.valence = valence;
         this.arousal = arousal;
